refactor: properties, exceptionhandler 패키지 이
