+++ src/main/java/me/hoyeon/shortlink/infrastructure/config/ErrorResponse.java
@@ -1,4 +1,4 @@
-package me.hoyeon.shortlink.config.exception;
+package me.hoyeon.shortlink.infrastructure.config;
 
 public record ErrorResponse(
     String code,
