+++ src/test/java/me/hoyeon/shortlink/integration/ui/RedirectControllerStandAloneSetupTest.java
@@ -9,7 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import me.hoyeon.shortlink.application.NotAccessibleUrlException;
 import me.hoyeon.shortlink.application.UrlNotFoundException;
 import me.hoyeon.shortlink.application.UrlShortenerService;
-import me.hoyeon.shortlink.config.exception.GlobalExceptionHandler;
+import me.hoyeon.shortlink.infrastructure.config.GlobalExceptionHandler;
 import me.hoyeon.shortlink.ui.RedirectController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
