+++ src/test/java/me/hoyeon/shortlink/unit/infrastructure/JavaEmailSenderTest.java
@@ -16,9 +16,9 @@ import jakarta.mail.internet.MimeMessage;
 import me.hoyeon.shortlink.application.EmailSendException;
 import me.hoyeon.shortlink.domain.EmailValidator;
 import me.hoyeon.shortlink.domain.InvalidEmailException;
-import me.hoyeon.shortlink.infrastructure.EmailProperties;
+import me.hoyeon.shortlink.infrastructure.config.EmailProperties;
 import me.hoyeon.shortlink.infrastructure.JavaEmailSender;
-import me.hoyeon.shortlink.infrastructure.RetryProperties;
+import me.hoyeon.shortlink.infrastructure.config.RetryProperties;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
